// src/utils/types.ts
export interface ServerInventory {
  id: string;
  hostname: string;
  ipAddress: string;
  environment: 'Prod' | 'Dev' | 'Test' | 'QA' | 'Lab' | 'Edu';
  description?: string;
  os: string;
  kernelPatchLevel?: string;
  cpu: string;
  memory: string;
  disk: string;
  javaVersion: string;
  jbossVersion?: string;
  webSphereVersion?: string;
  webLogicVersion?: string;
  apacheVersion?: string;
  nginxVersion?: string;
  tlsVersion?: string;
  ctgVersion?: string;
  hazelcastVersion?: string;
  mqVersion?: string;
  kafkaVersion?: string;
  redisVersion?: string;
  networkPorts?: string;
  owner?: string;
  lastUpdate?: string;
}